Commit before reorganising project structure

+++ indra_worker.py
@@ -47,6 +47,8 @@ def timeit(func, *args, **kwargs):
     return elapsed_time, result
 
 def get_statements_from_xmls():
+    logger.info(f"Worker {config['worker_id']}: Obtaining local statements from XMLs...")
+
     pkl_file = get_path('json_dir', 'get_statements_from_xmls.pkl')
 
     xml_files = os.listdir(config['xml_dir'])
@@ -57,8 +59,11 @@ def get_statements_from_xmls():
     end = (config['worker_id'] + 1) * chunk_size
 
     # Iterate over your chunk
-    logger.debug(f"Starting work over chunk [{start}, {end})...")
     times = []
+
+    end = min(end, len(xml_files))
+    chunk_size = end - start
+
     for i, xml_file in enumerate(xml_files[start:end]):
         xml_file = os.path.join(config['xml_dir'], xml_file)
 
@@ -75,7 +80,11 @@ def get_statements_from_xmls():
             
         # Periodically save progress
         if (i + 1 - start) % PICKLING_FREQUENCY == 0:
+            logger.info(f"Saving progress to pickle file...")
             pickle.dump((res, i), file=open(pkl_file, 'wb'))
+
+        progress_percentage = 100 * ((i + 1) / chunk_size)
+        logger.info(f"Progress: {i + 1}/{chunk_size} articles ({progress_percentage:.2f}%)")
 
     # Save statistics
     local_stats['processing_mean_time']  = mean(times)
@@ -92,7 +101,10 @@ def atomically_io(callback, wid=None):
         except Timeout: # Lock still under use
             pass
 
-def consolidate_stmts(stmts, pkl_prefix, out_filename):
+def consolidate_stmts(stmts, pkl_prefix, out_filename, master=False):
+    who = "MASTER" if master else f"Worker {config['worker_id']}" 
+    logger.info(f"{who}: Consolidating statements...")
+
     # Grounding
     pkl_file = get_path('pkl_dir', pkl_prefix + '_grounding.pkl')
     t, stmts = timeit(lambda: load_or_compute(pkl_file, lambda: ac.map_grounding(stmts, save=pkl_file, gilda_mode='local')))
@@ -117,55 +129,82 @@ def consolidate_stmts(stmts, pkl_prefix, out_filename):
     # TODO: pipe to a zip w/checksum...?
 
 def dump_local_stats():
+    logger.info(f"Worker {config['worker_id']}: Dumping statistics of local processing...")
+    
     local_stats_df = pd.DataFrame(local_stats, index=[0])
     local_stats_df['worker_id'] = config['worker_id']
+    
+    # Reorder columns to have 'worker_id' first
+    columns_order = ['worker_id'] + [col for col in local_stats_df.columns if col != 'worker_id']
+    local_stats_df = local_stats_df[columns_order]
 
     csv_path = get_path('csv_dir', 'local_stats.csv')
     atomically_io(lambda: local_stats_df.to_csv(csv_path, index=False))
 
 def aggregate_local_stats():
+    logger.info("MASTER: Aggregating statistics of workers' local processing...")
+    
     def try_read_csv(csv_path):
         if os.path.exists(csv_path) and os.stat(csv_path).st_size > 0:
             return pd.read_csv(csv_path, index_col=0)
         return None
 
-    is_active  = [False] * config['num_workers']
-    cnt_active = config['num_workers']
     worker_dfs = [None] * config['num_workers']
+    cnt_active = config['num_workers']
 
+    logger.info(f"MASTER: Initially {cnt_active} procs are active")
     while cnt_active > 0:
+        logger.info(f"MASTER: Still waiting for {cnt_active} procs...")
+
+        is_active = [True if worker_dfs[wid] is None else False for wid in range(config['num_workers'])]
+        logger.info(f"MASTER: is_active = {is_active}")
+
         for wid in range(config['num_workers']):
+            if worker_dfs[wid] is not None:
+                continue # Already processed this worker
+
             csv_path = get_path('csv_dir', 'local_stats.csv', wid)
             worker_dfs[wid] = atomically_io(lambda: try_read_csv(csv_path), wid)
             if worker_dfs[wid] is not None:
-                is_active[wid] = False
                 cnt_active -= 1
+                # Lock not needed anymore
+                lock_path = to_worker_fname('lock', wid)
+                os.remove(lock_path)
+        time.sleep(2)
     
-    # Remove locks (not needed anymore)
-    for wid in range(config['num_workers']):  
-        lock_path = to_worker_fname('lock', wid)
-        os.remove(lock_path)
+    logger.info(f"MASTER: Now only {cnt_active} are active (should be 0)")
+    is_active = [True if worker_dfs[wid] is None else False for wid in range(config['num_workers'])]
+    logger.info(f"MASTER: is_active = {is_active}")
 
-    return pd.concat(worker_dfs)
+    # Concatenate the workers' DataFrames
+    all_workers_df = pd.concat(worker_dfs)
 
-def make_final_stats(all_local_stats_df):
-    # Calculate the mean of each column
-    mean_values = all_local_stats_df.mean(numeric_only=True)
-
-    # Create a new DataFrame row summarizing the others
+    # Compute their mean values and append them as a new row
+    mean_values = all_workers_df.mean(numeric_only=True)
     new_row_id = 'total (mean)'
     mean_row = pd.DataFrame([mean_values], index=[new_row_id])
-    mean_row['worker_id'] =  new_row_id
+    mean_row['worker_id'] = new_row_id
 
-    # Dump the updated DataFrame
-    final_stats_df = pd.concat([mean_row, all_local_stats_df], ignore_index=False)
-    csv_path = get_path('csv_dir', 'final_stats.csv')
-    final_stats_df.to_csv(csv_path, index=True)
+    # Reorder columns to have 'worker_id' first
+    columns_order = ['worker_id'] + [col for col in all_workers_df.columns if col != 'worker_id']
+    mean_row = mean_row[columns_order]
 
-    return final_stats_df
+    return pd.concat([mean_row, all_workers_df], ignore_index=False)
 
+def dump_final_stats():
+    logger.info("MASTER: Dumping final statistics...")
+
+    aggregation_stats = {k: (v if k.startswith('consolidation') else '') for k, v in local_stats.items()}
+    aggregation_stats_df = pd.DataFrame(aggregation_stats, index=[0])
+    aggregation_stats_df['worker_id'] = 'aggregation'
+
+    final_stats_df = pd.concat([aggregated_stats_df, aggregation_stats_df], ignore_index=False)
+
+    csv_path = get_path('csv_dir', 'final_stats.csv', 'master')
+    final_stats_df.to_csv(csv_path, index=False)
 
 def get_stmts_from_jsons():
+    logger.info("MASTER: Extracting local statements from workers' jsons...")
     stmts = []
     for json in os.listdir(config['json_dir']):
         stmts += statements.stmts_from_json_file(os.path.join(config['json_dir'], json))
@@ -200,13 +239,20 @@ if __name__ == "__main__":
     ## 2. Consolidate local statements
     consolidate_stmts(local_stmts, 'consolidation_local', 'local_results.json')
 
+    if config['worker_id'] != 0:
+        time.sleep(45)
+
     ## 3. Dump local stats to CSV file
     dump_local_stats()
 
-    ## *4. Master-only: wait for workers to finish and aggregate their results
+    ## (Master-only)
     if config['worker_id'] == 0:
-        all_local_stats_df = aggregate_local_stats()
-        make_final_stats(all_local_stats_df)
-
+        ### 4. Wait for workers to finish, aggregate their statistics
+        aggregated_stats_df = aggregate_local_stats()
+        
+        ### 4. Combine workers' results
         all_stmts = get_stmts_from_jsons()
-        consolidate_stmts(all_stmts, 'consolidation_final', 'final_results.json') 
+        consolidate_stmts(all_stmts, 'consolidation_final', 'final_results.json')
+
+        ### 5. Dump aggregated statistics to CSV file
+        dump_final_stats()
